backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
PyPortfolioOpt>=1.5.0
python-multipart>=0.0.6
scipy>=1.10.0
pyjwt>=2.8.0
SQLAlchemy>=2.0.0
requests>=2.31.0
supabase>=2.0.0
git+https://github.com/druzsan/justetf-scraping.git
